+++ argument_parser.nim
@@ -1,7 +1,7 @@
-## Command line parsing module for Nimrod.
+## Command line parsing module for Nim.
 ##
-## `Nimrod <http://nimrod-code.org>`_ provides the `parseopt module
-## <http://nimrod-code.org/parseopt.html>`_ to parse options from the
+## The `Nim programming language <http://nim-lang.org>`_ provides the `parseopt
+## module <http://nim-lang.org/parseopt.html>`_ to parse options from the
 ## commandline. This module tries to provide functionality to prevent you from
 ## writing commandline parsing and let you concentrate on providing the best
 ## possible experience for your users.
@@ -13,18 +13,24 @@ import os, strutils, tables, math, parseutils, sequtils, sets, algorithm,
   unicode
 
 const
-  VERSION_STR* = "0.2.0" ## Module version as a string.
-  VERSION_INT* = (major: 0, minor: 2, maintenance: 0) ## \
+  version_int* = (major: 0, minor: 4, maintenance: 0) ## \
   ## Module version as an integer tuple.
   ##
   ## Major versions changes mean a break in API backwards compatibility, either
   ## through removal of symbols or modification of their purpose.
   ##
-  ## Minor version changes can add procs (and maybe default parameters). Minor
-  ## odd versions are development/git/unstable versions. Minor even versions
-  ## are public stable releases.
+  ## Minor version changes can add procs (and maybe default parameters).
   ##
   ## Maintenance version changes mean bugfixes or non API changes.
+  ##
+  ## Odd versions are development/git/unstable versions. Even versions are
+  ## public stable releases.
+
+  version_str* = ($version_int.major & "." & $version_int.minor & "." &
+    $version_int.maintenance) ## \
+  ## Module version as a string.
+  ##
+  ## The string has the format ``digits.digits.digits``.
 
 # - Types
 
@@ -73,12 +79,15 @@ type
     ## this field to write a generic proc to deal with parsed parameters, but
     ## nothing prevents you from accessing directly the type of field you want
     ## if you expect only one kind.
+    ##
+    ## To create objects of this type you can use the `new_parsed_parameter()
+    ## <#new_parsed_parameter>`_ template.
     case kind*: Tparam_kind
     of PK_EMPTY: nil
     of PK_INT: int_val*: int
-    of PK_BIGGEST_INT: big_int_val*: biggestInt
+    of PK_BIGGEST_INT: big_int_val*: BiggestInt
     of PK_FLOAT: float_val*: float
-    of PK_BIGGEST_FLOAT: big_float_val*: biggestFloat
+    of PK_BIGGEST_FLOAT: big_float_val*: BiggestFloat
     of PK_STRING: str_val*: string
     of PK_BOOL: bool_val*: bool
     of PK_HELP: nil
@@ -157,17 +166,15 @@ proc `$`*(data: Tparsed_parameter): string {.procvar.} =
 
 
 template new_parsed_parameter*(tkind: Tparam_kind, expr): Tparsed_parameter =
-  ## Handy compile time template to build Tparsed_parameter object variants.
-  ##
-  ## The problem with object variants is that you first have to initialise them
-  ## to a kind, then assign values to the correct variable, and it is a little
-  ## bit annoying.
+  ## Handy compile time template to build `Tparsed_parameter
+  ## <#Tparsed_parameter>`_ object variants.
   ##
   ## Through this template you specify as the first parameter the kind of the
-  ## Tparsed_parameter you want to build, and directly the value it will be
-  ## initialised with. The template figures out at compile time what field to
-  ## assign the variable to, and thus you reduce code clutter and may use this
-  ## to initialise single assignments variables in `let` blocks. Example:
+  ## `Tparsed_parameter <#Tparsed_parameter>`_ you want to build, and directly
+  ## the value it will be initialised with. The template figures out at compile
+  ## time what field to assign the variable to, and thus you reduce code
+  ## clutter and may use this to initialise single assignments variables in
+  ## `let` blocks. Example:
   ##
   ## .. code-block:: nimrod
   ##   let
@@ -176,6 +183,28 @@ template new_parsed_parameter*(tkind: Tparam_kind, expr): Tparsed_parameter =
   ##     # The following line doesn't compile due to
   ##     # type mismatch: got (string) but expected 'int'
   ##     #parsed_param3 = new_parsed_parameter(PK_INT, "231")
+  ##
+  ## At some point the Nim programming language introduced initialization of
+  ## object variants through a new constructor syntax where you specify the
+  ## names of the fields you want to initialise. The new constructor syntax is
+  ## slightly more verbose and you need to know which field of the
+  ## `Tparsed_parameter <#Tparsed_parameter>`_ you want to set, but it has the
+  ## advantage of working with runtime variables, something this template
+  ## doesn't support.  The constructor syntax looks like this:
+  ##
+  ## .. code-block:: nimrod
+  ##   let
+  ##     parsed_param1 = Tparsed_parameter(kind: PK_FLOAT,
+  ##       float_val: 3.41)
+  ##     parsed_param2 = Tparsed_parameter(kind: PK_BIGGEST_INT,
+  ##       big_int_val: 2358123 * 23123)
+  ##     parsed_param3 = Tparsed_parameter(kind: PK_INT, str_val: "231")
+  ##
+  ## Note however that this new constructor syntax won't catch the error where
+  ## ``parsed_param3`` is being assigned an incorrect field at compile time:
+  ## you will get an ``EInvalidField`` exception raised at runtime in debug
+  ## builds, and also in release builds if you compile with
+  ## ``--fieldChecks:on``.
   var result {.gensym.}: Tparsed_parameter
   result.kind = tkind
   when tkind == PK_EMPTY: nil
@@ -230,14 +259,15 @@ template run_custom_proc(parsed_parameter: Tparsed_parameter,
   ##
   ## Pass in the string of the parameter triggering the call. If the
   if not custom_validator.isNil:
+    try:
+      let message = custom_validator(parameter, parsed_parameter)
+      if not message.isNil and message.len > 0:
+        raise_or_quit(EInvalidValue, ("Failed to validate value for " &
+          "parameter $1:\n$2" % [escape(parameter), message]))
     except:
       raise_or_quit(EInvalidValue, ("Couldn't run custom proc for " &
         "parameter $1:\n$2" % [escape(parameter),
         getCurrentExceptionMsg()]))
-    let message = custom_validator(parameter, parsed_parameter)
-    if not message.isNil and message.len > 0:
-      raise_or_quit(EInvalidValue, ("Failed to validate value for " &
-        "parameter $1:\n$2" % [escape(parameter), message]))
 
 
 proc parse_parameter(quit_on_failure: bool, param, value: string,
@@ -292,9 +322,9 @@ proc parse_parameter(quit_on_failure: bool, param, value: string,
       raise_or_quit(EInvalidValue, ("parameter $1 requires a " &
         "float, but $2 can't be parsed into one") % [param, escape(value)])
   of PK_EMPTY:
-    nil
+    discard
   of PK_HELP:
-    nil
+    discard
 
 
 template build_specification_lookup():
@@ -359,7 +389,7 @@ proc parse*(expected: seq[Tparameter_specification] = @[],
   # Prepare the input parameter list, maybe get it from the OS if not available.
   var args = args
   if args == nil:
-    let total_params = ParamCount()
+    let total_params = paramCount()
     #echo "Got no explicit args, retrieving from OS. Count: ", total_params
     newSeq(args, total_params)
     for i in 0..total_params - 1:
@@ -484,11 +514,3 @@ proc echo_help*(expected: seq[Tparameter_specification] = @[],
   for line in build_help(expected,
       type_of_positional_parameters, bad_prefixes, end_of_options):
     echo line
-
-
-when isMainModule:
-  # Simply tests code embedded in docs.
-  let
-    parsed_param1 = new_parsed_parameter(PK_FLOAT, 3.41)
-    parsed_param2 = new_parsed_parameter(PK_BIGGEST_INT, 2358123 * 23123)
-    #parsed_param3 = new_parsed_parameter(PK_INT, "231")